Organize tests directory
Because the tests directory contains tests for a lot of different
scripts, it is necessary to create subdirectories for model, utilities,
and parsing.
