+++ tests/model/coord.test.js
@@ -1,6 +1,6 @@
-const Coord = require('../model/coord.js');
-const Port = require('../model/port.js');
-const Validation = require('../utils/validation.js');
+const Coord = require('../../model/coord.js');
+const Port = require('../../model/port.js');
+const Validation = require('../../utils/validation.js');
 
 // Suppress console logs
 console.log = jest.fn();
