+++ tests/model/port.test.js
@@ -1,6 +1,6 @@
-const Port = require('../model/port.js');
-const Coord = require('../model/coord.js');
-const Validation = require('../utils/validation.js');
+const Port = require('../../model/port.js');
+const Coord = require('../../model/coord.js');
+const Validation = require('../../utils/validation.js');
 
 //Suppress console logs
 console.log = jest.fn();
